requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Pillow>=11.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
